feat: novos endpoint para cadastro de student

+++ app/Services/CursoService.php
@@ -4,14 +4,17 @@ namespace App\Services;
 
 use App\DTO\Cursos\CreateCursoDTO;
 use App\DTO\Cursos\UpdateCursoDTO;
+use App\Models\Registration;
 use App\Repositories\CursoRepositoryInterface;
 use App\Repositories\PaginationInterface;
+use App\Repositories\Registrations\RegistrationRepositoryInterface;
 use stdClass;
 
 class CursoService
 {
     public function __construct(
         protected CursoRepositoryInterface $repository,
+        protected RegistrationRepositoryInterface $repositoryRegistration
     ) {
     }
 
@@ -40,8 +43,14 @@ class CursoService
         return $this->repository->update($dto);
     }
 
-    public function delete(string $id): void
+    public function delete(string $id): array
     {
+        $response = ['error' => ""];
+        $exist = $this->repositoryRegistration->isExistThisCurso($id);
+        if ($exist) {
+            $response["error"] = "Você nao pode excluir esse curso.";
+            return $response;
+        }
         $this->repository->delete($id);
     }
 }
